[spring-blog] - add new table 'Roles' and add logs

+++ spring-boot-blog/src/main/java/com/spring/blog/security/TokenProvider.java
@@ -37,13 +37,13 @@ public class TokenProvider {
                 .compact();
     }
 
-    public Integer getUserIdFromToken(String token) {
+    public Long getUserIdFromToken(String token) {
         Claims claims = Jwts.parser()
                 .setSigningKey(appProperties.getAuth().getTokenSecret())
                 .parseClaimsJws(token)
                 .getBody();
 
-        return Integer.parseInt(claims.getSubject());
+        return Long.parseLong(claims.getSubject());
     }
 
     public boolean validateToken(String authToken) {
